Break out game bars
